Rename `matcher` to `extractor`
This will allow us to add other matchers that drop measurements rather
than extract tags from them.
