+++ regexengine/matchers.go
@@ -1,0 +1,35 @@
+package regexengine
+
+type match struct {
+	name string
+	tags map[string]string
+}
+
+type matchers struct {
+	extractors []extractor
+}
+
+func (m matchers) ExtractTagsFromMetric(metric Measurement) *match {
+	result := &match{
+		name: metric.Name(),
+		tags: map[string]string{},
+	}
+	matched := false
+
+	for _, matcher := range m.extractors {
+		// Merge all results. Some rules will only modify part of a metric name
+		if r := matcher.ApplyTo(result.name); r != nil {
+			matched = true
+			result.name = r.name
+			for k, v := range r.tags {
+				result.tags[k] = v
+			}
+		}
+	}
+
+	if !matched {
+		return nil
+	}
+
+	return result
+}
