+++ regexengine/rules.go
@@ -16,12 +16,14 @@ var (
 // extracting tags from metric names and compiles them into regex patterns that
 // can be applied to metric names.
 func CompileRulesIntoMatchers(rules []dsl.Rule) matchers {
-	matches := make(matchers, 0, len(rules))
+	compiledMatchers := matchers{
+		extractors: []extractor{},
+	}
 
 	for _, r := range rules {
 		var patternPrefix, patternSuffix string
 
-		matcher := matcher{RequiredPrefix: r.RequiredPrefix, ReplaceWith: r.ReplaceWith}
+		matcher := extractor{RequiredPrefix: r.RequiredPrefix, ReplaceWith: r.ReplaceWith}
 
 		rule := r.PartialMatch
 		if r.CompleteMatch != "" {
@@ -36,10 +38,10 @@ func CompileRulesIntoMatchers(rules []dsl.Rule) matchers {
 			fmt.Sprintf("%s%s%s", patternPrefix, pattern, patternSuffix),
 		)
 
-		matches = append(matches, matcher)
+		compiledMatchers.extractors = append(compiledMatchers.extractors, matcher)
 	}
 
-	return matches
+	return compiledMatchers
 }
 
 func compileRuleIntoPattern(rule string, customPatterns map[string]*regexp.Regexp) string {
